+++ NEWS.md
@@ -1,6 +1,9 @@
 # macleish 0.3.1
 
 * Fixed Nick's name in DESCRIPTION and added URL and Bug Reports
+* Added maple sap data
+* Added `macleish_elevation()` to retrieve GIS from MassGIS
+* Synchronized with `etl` version 0.3.6
 
 # macleish 0.3.0